% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Get Parameters for QC-RLSC Normalization}
\usage{
get_params(
  omicsData,
  block_cname,
  qc_cname,
  qc_ind,
  order_cname,
  missing_thresh,
  rsd_thresh
)
}
\arguments{
\item{omicsData}{an omicsData object (metabData, lipidData, pepData, or
proData) created using the pmartR package, where any zero values have
already been replaced with NAs and a log transformation has already been
applied}

\item{block_cname}{character string giving name of column in omicsData$f_Data
that contains block (or batch) information}

\item{qc_cname}{character string giving name of column in omicsData$f_data
that contains the factor variable indicating whether sample is QC or not}

\item{qc_ind}{character string giving the value from the qc_cname column that
indicates a QC sample}

\item{order_cname}{character string giving name of column in omicsData$f_data
that contains the run order}

\item{missing_thresh}{numeric threshold, between 0 and 1, used for filtering
out biomolecules. See details for more information. A value of 0.5 is
reasonable.}

\item{rsd_thresh}{numeric threshold used for filtering metabolites. See
details for more information. A value of 0.3 is reasonable.}
}
\value{
list with elements final_ests (used as \code{optimal_params} input
 argument to \code{qcrlsc}) and bad_feats (features, or biomolecules, that
 need to be filtered out prior to using \code{qcrlsc})
}
\description{
Obtain parameters for quality control-based robust LOESS (locally estimated
scatterplot smoothing) signal correction
}
\details{
Use this function to get the optimal parameter values to use in a
 subsequent call to \code{qcrlsc}, as well as a list of biomolecules that
 should be removed from the dataset prior to QC-RLSC normalization
}
\references{
Dunn,W.B., Broadhurst,D., Begley,P., Zelena,E., Francis-McIntyre,S., Anderson,N., Brown,M.,
Knowles,J.D., Halsall,A., Haselden,J.N. et al. (2011) Procedures for
large-scale metabolic profiling of serum and plasma using gas chromatography and
liquid chromatography coupled to mass spectrometry. Nat. Protoc., 6, 1060-1083
}
\author{
Lisa Bramer, Kelly Stratton
}
