% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_qcrlsc.R
\name{normalize_qcrlsc}
\alias{normalize_qcrlsc}
\title{QC-RLSC Normalization}
\usage{
normalize_qcrlsc(
  omicsData_filt,
  optimal_params,
  block_cname,
  qc_cname,
  qc_ind,
  backtransform = FALSE
)
}
\arguments{
\item{omicsData_filt}{an omicsData object (metabData, lipidData,
pepData, or proData) created using the pmartR package, where any zero
values have already been replaced with NAs and a log transformation has
already been applied, where any biomolecules that were identified by
\code{get_params()} as needing to be removed have been removed}

\item{optimal_params}{final_ests element of the output from \code{get_params()} function (this is
needed for Span and Poly_Degree values)}

\item{block_cname}{character string giving name of column in omicsData_filt$f_Data
that contains block (or batch) information. Values in this column must be numeric.}

\item{qc_cname}{character string giving name of column in omicsData_filt$f_data
that contains the factor variable indicating whether sample is QC or not}

\item{qc_ind}{character string giving the value from the qc_cname column that
indicates a QC sample}

\item{backtransform}{logical value indicated whether or not to backtransform
the data to put the normalized data back on the same scale as the original
data. Defaults to FALSE. If TRUE, the median of the y-values of the loess
curve is added to the normalized value for each biomolecule for each batch.}
}
\value{
omicsData object of same class as omicsData_filt, where e_data contains the QC-RLSC normalized values
}
\description{
Quality control-based robust LOESS (locally estimated scatterplot smoothing)
signal correction
}
\details{
This function applies the QC-RLSC normalization to each batch of input data. Must use \code{get_params} function first in order to get optimal_params input for \code{qcrlsc}.
}
\references{
Dunn,W.B., Broadhurst,D., Begley,P., Zelena,E., Francis-McIntyre,S., Anderson,N., Brown,M.,
Knowles,J.D., Halsall,A., Haselden,J.N. et al. (2011) Procedures for
large-scale metabolic profiling of serum and plasma using gas chromatography and
liquid chromatography coupled to mass spectrometry. Nat. Protoc., 6, 1060-1083
}
\author{
Kelly Stratton, Lisa Bramer
}
