% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_span.R
\name{optimal_span}
\alias{optimal_span}
\title{Get optimal span value}
\usage{
optimal_span(qc_data, lam = c(1, 2))
}
\arguments{
\item{qc_data}{data.frame of only the QC samples from one block/batch, where
the first column gives the run order and the second column gives the
abundance value}

\item{lam}{vector of (numeric) possible polynomial degrees to test in the
loess fitting; defaults to 1,}
}
\value{
list containing elements for: Lambda, Alpha, MSE
}
\description{
Get optimal span value for QC-RLSC normalization
}
\details{
This internal function is called by the \code{get_params} function
}
\author{
Lisa Bramer, Kelly Stratton
}
